requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-decouple==2.1
django-environ==0.9.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-parler==2.3
django-rosetta==0.9.8
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
polib==1.1.1
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
